32 bit Chacha
Greatly increase 32 bit Chacha speed. This is done by optimizing the number of loads and stores in the inner Chacha loop
Fix an issue with the IV's in the Chacha test. The IV's are 8 bytes long but the Chacha nonce is 12 bytes long. The wc_Chacha_SetIV function copies 12 bytes of input IV but with optimized tests it was copying 8 bytes of IV and 4 bytes of garbage.